Task 13 Fixed vk id type

+++ back-end/src/main/java/com/eltech/sh/model/Person.java
@@ -21,7 +21,7 @@ public class Person {
 
     @JsonProperty("id")
     @Index(unique=true,primary = true)
-    private Long vkId;
+    private Integer vkId;
 
     @JsonProperty("first_name")
     private String firstName;
@@ -34,17 +34,17 @@ public class Person {
     public Person() {
     }
 
-    public Person(Long vkId, String firstName, String lastName) {
+    public Person(Integer vkId, String firstName, String lastName) {
         this.vkId = vkId;
         this.firstName = firstName;
         this.lastName = lastName;
     }
 
-    public Long getVkId() {
+    public Integer getVkId() {
         return vkId;
     }
 
-    public void setVkId(Long vkId) {
+    public void setVkId(Integer vkId) {
         this.vkId = vkId;
     }
